clean up forms
