--- webook/arrangement/forms/requisition_service_form.py
@@ -1,6 +1,0 @@
-from django import forms
-from webook.arrangement.models import LooseServiceRequisition
-
-
-class RequisitionServiceForm (forms.Form):
-    pass
