--- webook/arrangement/forms/upload_files_to_arrangement_form.py
@@ -1,7 +1,0 @@
-from django import forms
-from webook.arrangement.models import Arrangement
-
-
-class UploadFilesToArrangementForm(forms.Form):
-    arrangement_slug = forms.SlugField()
-    file_field = forms.FileField(widget=forms.ClearableFileInput(attrs={'multiple': True}))
