--- webook/arrangement/forms/upload_files_to_event_serie_dialog.py
@@ -1,7 +1,0 @@
-from django import forms
-from webook.arrangement.models import Arrangement
-
-
-class UploadFilesToEventSerieForm(forms.Form):
-    event_serie_pk = forms.SlugField()
-    file_field = forms.FileField(widget=forms.ClearableFileInput(attrs={'multiple': True}))
