+++ webook/arrangement/models.py
@@ -1004,6 +1004,9 @@ class Event(TimeStampedModel, ModelTicketCodeMixin, ModelVisitorsMixin, ModelArc
     display_text = models.CharField(verbose_name=_("Screen Display Text"), max_length=255, blank=True, null=True)
     display_text_en = models.CharField(verbose_name=_("Screen Display Text(English)"), max_length=255, blank=True, null=True)
 
+    meeting_place = models.CharField(verbose_name=_("Meeting Place"), max_length=255, blank=True, null=True)
+    meeting_place_en = models.CharField(verbose_name=_("Meeting Place (English)"), max_length=255, blank=True, null=True)
+
     @property
     def is_buffer_event(self) -> bool:
         """Returns a bool indicating if this event is a buffer event for another event or not"""
